// frontend/src/api/model/declarationResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Apicius API
 * OpenAPI spec version: 1.0.0
 */
import type { CoreType } from './coreType';
import type { Refinement } from './refinement';

export interface DeclarationResponse {
  name?: string;
  coreType?: CoreType;
  refinement?: Refinement;
  oneOfValues?: string[];
  required?: boolean;
  description?: string;
}